+++ Cargo.toml
@@ -14,7 +14,6 @@ criterion = { version = "0.3", features = ["html_reports"] }
 rand_xorshift = "0.3"
 ark-std = { version = "0.3" }
 bincode = "1.3.3"
-halo2_proofs = { git = "https://github.com/privacy-scaling-explorations/halo2.git", rev="7a21656" }
 
 [dependencies]
 subtle = "2.4"
@@ -77,8 +76,4 @@ harness = false
 
 [[bench]]
 name = "msm"
-harness = false
-
-[[bench]]
-name = "msm-alt"
 harness = false
